requirements.txt
Flask==2.3.3
Pillow==10.0.1
psutil==5.9.0
gpiozero==1.6.2
spidev; platform_machine == "aarch64"
RPi.GPIO; platform_machine == "aarch64"